+++ backend/src/main/java/com/chestnut/backend/avatar/entity/Avatar.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class Avatar {
     @Id
     @Column(columnDefinition = "tinyint", nullable = false)
-    private Integer avatarId;
+    private Byte avatarId;
 
     @Column(columnDefinition = "varchar(10)", nullable = false)
     private String avatarName;
@@ -22,7 +22,7 @@ public class Avatar {
     private String avatarThumbnailUrl;
 
     @Column(columnDefinition = "tinyint", nullable = false)
-    private int lowerLimit;
+    private Byte lowerLimit;
 
     protected Avatar() {}
 }
